remove the single server's __del__
this was introducing a bug in the test cases where the mocks had been
tore down when the object was destroyed. This was not visible when the
test case ran fast enough (in most cases). However, sometimes the
destructor was called and shut down the system mpd (which is somewhat
annoying). So, no __del__ for the server (or anyone else, for that
matter)