remove the somewhat stupid root check and absolute path obsession
The root existence check is done anyways when trying to construct the
basepath -- no need to do that explicitly (and thus *twice*), and the
test case complexity is reduced. Also, relative paths are (if we don't
change the current directory) totally equivalent to absolute paths -- no
need to enforce absolute paths.