implement the player stubs
However, the property passthrough involves some ugly hack: to determine
whether to pick from the player or the current client, we have to
maintain a list of variable names. This could be worked around by having
a guard value (like __initialized) and checking __dict__ whether that
exists: if not, we just extend the player, else we don't allow new
attribute creation.