fix the bug introduced in 017b940b242909dd92b1513cafeb43cde9c93fd1
Which would not allow instantiating the player class. Also, we didn't
seem to have run the test suite, which also broke. The reason for this
was that the outsourced startup code needed the `prebuffering` property,
which depended on `_prebuffering`, which was set only *after* the
startup code. Thus, changing the call order (setting before calling
`start()`) allows proper initialization.