extend the playlist extraction to also understand pls files
which, I don't know why, have a hard to find spec (i.e. I couldn't find
it). So, this is just filtering lines that begin with "File", splitting
these by "=" and using the second part. Hopefully that is a valid
url (because that is currently not checked). Oh, and it also reduces
the test coverage. However, as increasing it again will definitly
require rewriting the current tests, it will be postponed.