add an update facility for updating property with adc values
however, I'm not really content with the names. It should probably named
something like "update_continuously_value_property" or
"continous_property" or something, and the adc attribute should be named
something like value factory. Also, the player should be something like
"object" because it's not really required to be a player (any object
could be used).