implement a rough version of the adc reader
the tests, however, are a different thing: as I want to be able to test
in absence of the gpiozero library (which depends on RPi.GPIO and thus
supposedly only works on a Raspberry Pi), I have to provide a fake.
Then, as the reader inherits from a gpiozero class, some complex monkey
patching is required: patch the base class. Instead (and perhaps really
easier and much more reliable), we could also just patch the methods and
attributes that are used by the test. Maybe a task for tomorrow?